Extracting Address class into main namespace
