+++ lib/street_address.rb
@@ -1,2 +1,3 @@
 require 'street_address/version'
+require 'street_address/address'
 require 'street_address/us'
